Moved test to a seperate file
